move: delivery manager application 패키지 구조 변경
