+++ delivery/src/main/java/com/nowayback/delivery/application/deliverymanager/exception/DeliveryManagerApplicationException.java
@@ -1,4 +1,4 @@
-package com.nowayback.delivery.application.exception;
+package com.nowayback.delivery.application.deliverymanager.exception;
 
 import com.nowayback.common.exception.GlobalException;
