+++ delivery/src/test/java/com/nowayback/delivery/presentation/DeliveryManagerControllerTest.java
@@ -2,8 +2,8 @@ package com.nowayback.delivery.presentation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nowayback.common.security.annotation.UserRole;
-import com.nowayback.delivery.application.DeliveryManagerService;
-import com.nowayback.delivery.application.dto.DeliveryManagerResult;
+import com.nowayback.delivery.application.deliverymanager.DeliveryManagerService;
+import com.nowayback.delivery.application.deliverymanager.dto.DeliveryManagerResult;
 import com.nowayback.delivery.presentation.dto.request.CreateDeliveryManagerRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
